Consolidated the Menus and used Inheritance for products

+++ capstone/src/main/java/com/techelevator/view/Money.java
@@ -1,19 +1,21 @@
 package com.techelevator.view;
 
+import java.math.BigDecimal;
+
 public enum Money {
-    ONE_DOLLARS(1.00),
-    TWO_DOLLARS(2.00),
-    FIVE_DOLLARS(5.00),
-    TEN_DOLLARS(10.00);
+    ONE_DOLLARS(BigDecimal.valueOf(1.00)),
+    TWO_DOLLARS(BigDecimal.valueOf(2.00)),
+    FIVE_DOLLARS(BigDecimal.valueOf(5.00)),
+    TEN_DOLLARS(BigDecimal.valueOf(10.00));
 
 
-    private double value;
+    private BigDecimal value;
 
-    private Money(double value){
+    private Money(BigDecimal value){
         this.value = value;
     }
 
-    public double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 }
